+++ packages/ui/src/components/AccordionJsonView/index.tsx
@@ -4,15 +4,24 @@ import AccordionSummary from '@mui/material/AccordionSummary';
 import AccordionDetails from '@mui/material/AccordionDetails';
 import Typography from '@mui/material/Typography';
 import ExpandMoreIcon from '@mui/icons-material/ExpandMore';
-import SimpleJsonView from '../SimpleJsonView';
+import JsonTreeView from '../JsonTreeView';
 
 type TProps = {
   header: string;
   textClassName?: string;
   defaultExpanded?: boolean;
+  filterBasePath?: string;
+  onFilterAdded?: (expression: string) => void;
 };
 const AccordionJsonView: React.FC<TProps> = (props) => {
-  const { children, header, textClassName, defaultExpanded = true } = props;
+  const {
+    children,
+    header,
+    textClassName,
+    defaultExpanded = true,
+    filterBasePath,
+    onFilterAdded,
+  } = props;
   return (
     <div>
       <Accordion defaultExpanded={defaultExpanded}>
@@ -20,7 +29,13 @@ const AccordionJsonView: React.FC<TProps> = (props) => {
           <Typography>{header}</Typography>
         </AccordionSummary>
         <AccordionDetails>
-          <SimpleJsonView className={textClassName}>{children}</SimpleJsonView>
+          <JsonTreeView
+            className={textClassName}
+            filterBasePath={filterBasePath}
+            onFilterAdded={onFilterAdded}
+          >
+            {children as string}
+          </JsonTreeView>
         </AccordionDetails>
       </Accordion>
     </div>