+++ packages/ui/src/network/queries/get-queues.ts
@@ -12,6 +12,7 @@ export const getQueues = (): Promise<GetQueuesQuery> =>
           readonly
           name
           keyPrefix
+          group
           isPaused
           jobsCounts {
             waiting